feat: remove disused discovery cache
